+++ src/Products/EasyNewsletter/views/newsletter_issue_send.py
@@ -18,14 +18,21 @@ import transaction
 log = logging.getLogger('Products.EasyNewsletter')
 
 
-class PloneMessageSendMixin():
+class PloneMessageSendMixin:
     """
     """
+
     def __init__(self):
         pass
 
 
-class Message(PloneMessageSendMixin, emails.message.MessageTransformerMixin, emails.message.MessageSignMixin, emails.message.MessageBuildMixin, emails.message.BaseMessage):
+class Message(
+    PloneMessageSendMixin,
+    emails.message.MessageTransformerMixin,
+    emails.message.MessageSignMixin,
+    emails.message.MessageBuildMixin,
+    emails.message.BaseMessage,
+):
     """
     Email message with:
     - DKIM signer
@@ -121,7 +128,9 @@ class NewsletterIssueSend(BrowserView):
         # get issue data
         issue_data = issue_data_fetcher()
         for receiver in receivers:
-            personalized_html = issue_data_fetcher.personalize(receiver, issue_data['body_html'])
+            personalized_html = issue_data_fetcher.personalize(
+                receiver, issue_data['body_html']
+            )
             # get plain text version
             personalized_plaintext = issue_data_fetcher.create_plaintext_message(
                 personalized_html
@@ -138,18 +147,20 @@ class NewsletterIssueSend(BrowserView):
             if 'HTTPLoaderError' in m.as_string():
                 log.exception(u"Transform message failed: {0}".format(m.as_string()))
             try:
-                self.mail_host.send(m.as_string())
+                self.mail_host.send(m.as_string(), immediate=True)
                 log.info('Send newsletter to "%s"' % receiver['email'])
                 send_counter += 1
             except Exception as e:  # noqa
                 log.exception(
                     'Sending newsletter to "%s" failed, with error "%s"!'
-                    % (receiver['email'], e))
+                    % (receiver['email'], e)
+                )
                 send_error_counter += 1
 
         log.info(
             'Newsletter was sent to (%s) receivers. (%s) errors occurred!'
-            % (send_counter, send_error_counter))
+            % (send_counter, send_error_counter)
+        )
 
         # change status only for a 'regular' send operation (not 'is_test')
         if not self.is_test:
@@ -158,6 +169,22 @@ class NewsletterIssueSend(BrowserView):
             self.request['enlwf_guard'] = False
             self.context.setEffectiveDate(DateTime())
             self.context.reindexObject(idxs=['effective'])
+            msg_type = "info"
+            additional_warning = ""
+            if send_error_counter:
+                msg_type = "warn"
+                additional_warning = _(
+                    "\nPlease check the log files, for more details!"
+                )
+            api.portal.show_message(
+                message=_(
+                    'Newsletter was sent to ({0}) receivers. ({1}) errors occurred!{2}'.format(
+                        send_counter, send_error_counter, additional_warning
+                    )
+                ),
+                request=self.request,
+                type=msg_type,
+            )
 
     @property
     def salutation_mappings(self):
@@ -197,12 +224,14 @@ class NewsletterIssueSend(BrowserView):
             if test_receiver == "":
                 test_receiver = enl.test_email
             salutation = salutation_mappings.get('default', '')
-            receivers = [{
-                'email': test_receiver,
-                'fullname': 'Test Member',
-                'salutation': salutation.get(self.context.language, ''),
-                # 'nl_language': self.language
-            }]
+            receivers = [
+                {
+                    'email': test_receiver,
+                    'fullname': 'Test Member',
+                    'salutation': salutation.get(self.context.language, ''),
+                    # 'nl_language': self.language
+                }
+            ]
             return receivers
 
         # only send to all subscribers if the exclude all subscribers
@@ -210,7 +239,9 @@ class NewsletterIssueSend(BrowserView):
         # get Subscribers
         enl_receivers = []
         if not self.context.exclude_all_subscribers:
-            for subscriber_brain in api.content.find(portal_type='Newsletter Subscriber', context=enl):
+            for subscriber_brain in api.content.find(
+                portal_type='Newsletter Subscriber', context=enl
+            ):
                 if not subscriber_brain:
                     continue
                 subscriber = subscriber_brain.getObject()
@@ -222,11 +253,12 @@ class NewsletterIssueSend(BrowserView):
                     'name_prefix': subscriber.name_prefix,
                     'firstname': subscriber.firstname or u'',
                     'lastname': subscriber.lastname or u'',
-                    'fullname': ' '.join([subscriber.firstname or u'',
-                                          subscriber.lastname or u'']),
+                    'fullname': ' '.join(
+                        [subscriber.firstname or u'', subscriber.lastname or u'']
+                    ),
                     'salutation': salutation.get(
                         None,  # subscriber.getNl_language(),
-                        salutation.get(self.context.language or 'en', '')
+                        salutation.get(self.context.language or 'en', ''),
                     ),
                     'uid': subscriber.UID(),
                     # 'nl_language': subscriber.getNl_language()